requirements.txt
# Core components
streamlit==1.37.0
requests==2.31.0
python-dotenv==1.0.1

# Envoy AI Gateway requirements
pyyaml==6.0.1
grpcio==1.62.1
protobuf==4.25.3

# Google ADK
google-adk==0.5.0
vertexai==1.45.0
google-cloud-aiplatform==1.49.0

# LLM Provider SDKs
openai==1.18.0
anthropic==0.21.0

# Visualization and UI
plotly==5.25.0
pandas==2.2.1
numpy==1.26.4
altair==5.2.0

# Tools for financial calculations
numpy-financial==1.0.0

# Testing
pytest==8.0.0
